+++ source/core/rendering/rendering_driver.hpp
@@ -44,7 +44,7 @@ private:
 						 progress::Sink& progressor);
 
 	size_t calculate_progress_range(const scene::Scene& scene,
-									const scene::camera::Camera& camera, size_t num_tiles) const;
+									const scene::camera::Camera& camera, uint32_t num_tiles) const;
 
 	std::shared_ptr<Surface_integrator_factory> surface_integrator_factory_;
 	std::shared_ptr<sampler::Sampler> sampler_;
